Prediction Cross-validation

Function to calculate predictive powers of different models

Usage

Prediction_cross_validation(
  person,
  k = 10,
  threshold = 0,
  disease,
  method = "GWAS",
  liabilities = person$FAM$Status
)

Arguments

person

A .rds file with an FBM.code256 and accompanying FAM and MAP.

k

Number of folds to be used in cross-validation. Default is 10.

threshold

Vector of P-values to be used in thresholding. Default does not use thresholding.

disease

List with properties of disease.

method

Method to use for prediction. Possible methods are "GWAS", "GWAX", "LTFH". Default is "GWAS".

liabilities

Vector of liabilities used for prediction with "LTFH" method. If not specified, uses "GWAS" method instead.

Value

A list with 3 values: a tibble with average and best scores for each threshold, the best score, and the best model.